# docs/swagger.yaml
basePath: /api/v1
...
host: localhost:3000
info:
  contact:
    email: support@split-go.com
    name: Split Go API Support
    url: http://www.split-go.com/support
  description: 分帳系統 API 文檔 - 支援群組管理、交易記錄、結算功能
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Split Go API
  version: "1.0"
paths:
...
  /groups/{id}/balance:
    get:
      description: 獲取群組中每個成員的總欠情況
      parameters:
      - description: 群組ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 群組總欠情況
          schema:
            properties:
              data:
                items:
                  properties:
                    name:
                      type: string
                    net_balance:
                      type: number
                    user_id:
                      type: integer
                    username:
                      type: string
                  type: object
                type: array
              error:
                type: boolean
            type: object
        "401":
          description: 未授權
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
        "403":
          description: 不是群組成員
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
        "404":
          description: 群組不存在
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
        "500":
          description: 服務器內部錯誤
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: 獲取群組總欠
      tags:
      - 交易
...
  /groups/{id}/transactions:
    get:
      description: 獲取指定群組的所有交易記錄
      parameters:
      - description: 群組ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 群組交易列表
          schema:
            properties:
              data:
                items:
                  properties:
                    amount:
                      type: number
                    category:
                      properties:
                        id:
                          type: integer
                        name:
                          type: string
                      type: object
                    category_id:
                      type: integer
                    created_at:
                      type: string
                    creator:
                      properties:
                        id:
                          type: integer
                        name:
                          type: string
                        username:
                          type: string
                      type: object
                    description:
                      type: string
                    group:
                      properties:
                        id:
                          type: integer
                        name:
                          type: string
                      type: object
                    group_id:
                      type: integer
                    id:
                      type: integer
                    paid_by:
                      type: integer
                    payer:
                      properties:
                        id:
                          type: integer
                        name:
                          type: string
                        username:
                          type: string
                      type: object
                    receipt_url:
                      type: string
                    split_type:
                      type: string
                    splits:
                      items:
                        properties:
                          amount:
                            type: number
                          id:
                            type: integer
                          percentage:
                            type: number
                          user:
                            properties:
                              id:
                                type: integer
                              name:
                                type: string
                              username:
                                type: string
                            type: object
                          user_id:
                            type: integer
                        type: object
                      type: array
                    updated_at:
                      type: string
                  type: object
                type: array
              error:
                type: boolean
            type: object
        "401":
          description: 未授權
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
        "403":
          description: 不是群組成員
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
        "404":
          description: 群組不存在
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
        "500":
          description: 服務器內部錯誤
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: 獲取群組交易列表
      tags:
      - 交易
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 在 header 中加入 "Bearer <token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"